+++ grace-shell/src/main/resources/META-INF/services/org.grails.cli.commands.Command
@@ -1,4 +1,5 @@
 org.grails.cli.command.generate.GenerateCommand
+org.grails.cli.command.gradle.GradleCommand
 org.grails.cli.command.destroy.DestroyCommand
 org.grails.cli.command.run.RunCommand
 org.grails.cli.commands.ApplicationTemplateCommand
@@ -11,4 +12,3 @@ org.grails.cli.commands.ListPluginsCommand
 org.grails.cli.commands.ListProfilesCommand
 org.grails.cli.commands.PluginInfoCommand
 org.grails.cli.commands.ProfileInfoCommand
-org.grails.cli.gradle.commands.GradleCommand
